pal: Fix compilation issues with new component
Fix compilation issues in the new component
upgrade.

+++ device/src/Bluetooth.cpp
@@ -1966,10 +1966,10 @@ void BtSco::convertCodecInfo(audio_lc3_codec_cfg_t &lc3CodecInfo,
         streamMapStr = match.suffix().str();
     }
 
-    PAL_DBG(LOG_TAG, "stream map out size: %d, stream map in size: %d", steamMapOut.size(), steamMapIn.size());
+    PAL_DBG(LOG_TAG, "stream map out size: %d, stream map in size: %d", (int)steamMapOut.size(), (int)steamMapIn.size());
     if ((steamMapOut.size() == 0) || (steamMapIn.size() == 0)) {
         PAL_ERR(LOG_TAG, "invalid size steamMapOut.size %d, steamMapIn.size %d",
-                steamMapOut.size(), steamMapIn.size());
+                (int)steamMapOut.size(), (int)steamMapIn.size());
         return;
     }
