+++ session/src/SessionAgm.cpp
@@ -560,7 +560,7 @@ int SessionAgm::write(Stream *s, int tag __unused, struct pal_buffer *buf, int *
     status = agm_session_write_with_metadata(agmSessHandle, &agm_buffer, &bytes_written);
 
     PAL_VERBOSE(LOG_TAG, "writing buffer (%zu bytes) to agmSessHandle device returned %d",
-             buf->size, bytes_written);
+             buf->size, (int)bytes_written);
 
     if (size)
         *size = bytes_written;
